+++ CHANGELOG.md
@@ -8,6 +8,9 @@ All notable changes to this project will be documented in this file. This projec
 - Dependency `symfony/psr-http-message-bridge` now allows `^1.0` as version 1 has now been released. Version `0.2.*` is
 still allowed to maintain backwards compatibility but will be removed in `v0.6` of this package.
 
+### Fixed
+- Amended testing that a model has been trashed so that it is compatible with Laravel 5.2 and 5.3.
+
 ## [0.5.0] - 2016-10-21
 
 This release is the first release to support Laravel 5.3.